Move TensorOptions to ATen/core
In preparation for adding DimnameList (which is in ATen) to
TensorOptions.
In the long term we'll move DimnameList to c10 because that is important
for named tensor performance (TensorImpl holds a DimnameList but right
now there is one layer of indirection to support it). However, for now,
since TensorOptions isn't used by c10, move it back into ATen/core.

Test Plan:
- wait for ci

gh-metadata: pytorch pytorch 22020 gh/zou3519/57/head